requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI APIs
anthropic==0.7.8
openai==1.3.7

# HTTP requests
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
structlog==23.2.0

# CORS and middleware
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1 